Getting started
I’ve removed some of the skeleton css that I don’t think we need, if
you want to add it back in then do so :)

I’ve set up a max width of 768px on our index.html - for portrait
tablet size. If this isn’t the right max width then feel free to update.